+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CRISPIN development version
 
+- Improve help message for `crispin run`. (#65, @kelly-sovacool)
+  - the default `--mode` is now `slurm`.
+
 ## CRISPIN 1.2.0
 
 - CRISPIN now depends on ccbr_tools v0.4 for updated jobby & spooker utilities. (#63, @kelly-sovacool)